Move useVisitors query to query file
